parentheses and unary operators
this op stack thing has become quite involved with edge cases, but this
seems to work. It would be nice to explicitly track the parentheses and
things, rather than presenting cryptic 'op stack was not empty'
messages, and generally to make this logic less spaghetti-like, if that
is even possible.

maybe this is what solving complicated problems is like though, and
maybe I have to get used to it a little bit!